move `check_expected_wheels` into `wheel_utils.py`
